Renamed scripts for easier access, edited readme md to reflect updates, continued cleaning my analysis nb
