+++ 5311/readme_analysis.md
@@ -3,12 +3,11 @@
 ## 5311
 
 ### Data Creation and Utils
-1. [data_prep](./data_prep.py): functions for cleaning our 3 data sources and merging them together. 
+1. [data_prep](./_data_prep.py): functions for cleaning our 3 data sources, merging, and aggregating them together. 
 2. [utils](./_utils.py): functions for charts and aggregating the merged data frames. 
-3. [add_inflation](./5311/add_inflation.ipynb)
-4. [add_rtpa.ipynb](./5311/add_rtpa.ipynb)
-5. [merge_crosswalk.ipynb](./5311/merge_crosswalk.ipynb)
-6. []()
+3. [agency crosswalk](./_agency_crosswalk.py): a crosswalk created to match agency names against the data sources, when the Cal ITP/NTD ID were not available. 
+4. [add_inflation](./5311/add_inflation.ipynb)
+5. [add_rtpa.ipynb](./5311/add_rtpa.ipynb)
 
 ### Analyses
 1. [5311.ipynb](./5311/5311.ipynb): 5311 organizations first pass
